requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp_retry==2.4.6
dataclasses_json==0.5.7
marshmallow==3.15.0
numpy==1.22.2
olefile==0.46
pandas==1.4.2
pdfplumber==0.6.0
python-dotenv==0.20.0
python_dateutil==2.8.2
python_magic==0.4.25
requests==2.27.1
tqdm==4.62.3
